+++ poem-openapi/src/types/any.rs
@@ -51,3 +51,41 @@ impl<T: Serialize + Send + Sync> ToJSON for Any<T> {
         serde_json::to_value(&self.0).unwrap_or_default()
     }
 }
+
+impl Type for Value {
+    const IS_REQUIRED: bool = true;
+
+    type RawValueType = Self;
+
+    type RawElementValueType = Self;
+
+    fn name() -> Cow<'static, str> {
+        "any".into()
+    }
+
+    fn schema_ref() -> MetaSchemaRef {
+        MetaSchemaRef::Inline(Box::new(MetaSchema::ANY))
+    }
+
+    fn as_raw_value(&self) -> Option<&Self::RawValueType> {
+        Some(self)
+    }
+
+    fn raw_element_iter<'a>(
+        &'a self,
+    ) -> Box<dyn Iterator<Item = &'a Self::RawElementValueType> + 'a> {
+        Box::new(self.as_raw_value().into_iter())
+    }
+}
+
+impl ParseFromJSON for Value {
+    fn parse_from_json(value: Value) -> ParseResult<Self> {
+        Ok(value)
+    }
+}
+
+impl ToJSON for Value {
+    fn to_json(&self) -> Value {
+        self.clone()
+    }
+}